requirements.txt
# Web Framework & Server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Core AI & Deep Learning Libraries
# Using CUDA 11.8. For other versions, check PyTorch official website.
# Ensure your NVIDIA driver is compatible.
#torch==2.3.0
#torchvision==0.18.0
#torchaudio==2.3.0
# If you have CUDA installed and want to specify the version:
# torch==2.3.0+cu121
# torchvision==0.18.0+cu121
# torchaudio==2.3.0+cu121
# command use : pip install torch torchvision torchaudio --index-url http://download.pytorch.org/whl/cu118
# 但是我们是用CUDA121，所以从118换去121
# Machine Learning & Scientific Computing
scikit-learn==1.4.2
scikit-fuzzy==0.4.2
numpy==1.26.4
pandas==2.2.2

# Image Processing
Pillow==10.3.0
opencv-python-headless==4.9.0.80

# Type Hinting & Data Validation (used by FastAPI)
pydantic==2.7.1

# Utilities
# tqdm for progress bars during training (optional but recommended)
tqdm==4.66.4
# matplotlib for data visualization (optional, useful for debugging)
matplotlib==3.8.4

PyYAML==6.0.1
httpx==0.27.0
#fasttext==0.9.2
python-jose==3.3.0
loguru
albumentations
#simple_image_download==0.5
imagehash==4.3.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
selenium==4.21.0
webdriver-manager
kaggle
roboflow